Check files are older than XX minutes (default 60) before deleting them.

+++ EventFilter/StorageManager/scripts/operations/smCleanupFiles.pl
@@ -1,12 +1,13 @@
-#!/usr/bin/env perl -w
-# $Id: smCleanupFiles.pl,v 1.1 2008/11/13 08:29:22 loizides Exp $
+#!/usr/bin/env perl
+# $Id: smCleanupFiles.pl,v 1.2 2009/03/03 13:43:43 jserrano Exp $
 
 use strict;
+use warnings;
 use DBI;
 use Getopt::Long;
 use File::Basename;
 
-my ($help, $debug, $nothing, $force, $execute, $maxfiles);
+my ($help, $debug, $nothing, $force, $execute, $maxfiles, $fileagemin);
 my ($hostname, $filename, $dataset, $stream, $config);
 my ($runnumber, $uptorun, $safety, $rmexitcode, $chmodexitcode, );
 my ($constraint_runnumber, $constraint_uptorun, $constraint_filename, $constraint_hostname, $constraint_dataset);
@@ -61,6 +62,7 @@ $hostname   = '';
 $rmexitcode = 0;
 $execute    = 1;
 $maxfiles   = 1;
+$fileagemin = 60;
 $force      = 0;
 $config     = "/opt/injectworker/.db.conf";
 
@@ -80,7 +82,8 @@ GetOptions(
 	   "filename=s"	   => \$filename,
 	   "dataset=s"	   => \$dataset,
            "stream=s"      => \$stream,
-           "maxfiles=i"    => \$maxfiles
+           "maxfiles=i"    => \$maxfiles,
+           "fileagemin=i"  => \$fileagemin
 	  );
 
 $help && usage;
@@ -157,16 +160,19 @@ while ( $nFiles<$maxfiles &&  (@row = $sth->fetchrow_array) ) {
     # remove file
     my $CHMODCOMMAND = "sudo chmod 666 $row[0]/$row[1]";
     my $RMCOMMAND    = "rm -f $row[0]/$row[1]";
-    $debug   && print "$RMCOMMAND \n";;
+    my $FILEAGEMIN   =  (time - (stat("$row[0]/$row[1]"))[9])/60;
+
+    $debug   && print "$FILEAGEMIN $fileagemin\n";
+    $debug   && print "$RMCOMMAND \n";
 
     $rmexitcode = 9998;
 
-    if ($execute && -e "$row[0]/$row[1]")
+    if ($execute && -e "$row[0]/$row[1]" && $FILEAGEMIN > $fileagemin)
     {
 	$chmodexitcode = system($CHMODCOMMAND);
 	$rmexitcode    = system($RMCOMMAND);	
 	$debug  && print "   ===> rm dat file successful?: $rmexitcode \n";
-    } elsif (!$execute && -e "$row[0]/$row[1]") {
+    } elsif (!$execute && -e "$row[0]/$row[1]" && $FILEAGEMIN > $fileagemin) {
 	#if we're not executing anything want to fake 
 	print "Pretending to remove $row[0]/$row[1] \n";
 	$rmexitcode = 0;
@@ -174,12 +180,15 @@ while ( $nFiles<$maxfiles &&  (@row = $sth->fetchrow_array) ) {
         if ($force) {
             print "File $row[0]/$row[1] does not exist, but force=1, so continue \n";
             $rmexitcode = 0;
+        } elsif ($FILEAGEMIN < $fileagemin)  {
+            print "File $row[0]/$row[1] too young to die\n";
         } else {
             print "File $row[0]/$row[1] does not exist \n";
+            $rmexitcode =0;
         }
     }
     $debug && print "\n";
-    $rmexitcode =0;
+    #$rmexitcode =0;
     # check file was really removed
     if ($rmexitcode != 0)
     {
